Separate logging of HTTP in and out
